+++ lab3/scalar.c
@@ -3,8 +3,8 @@
 #include <locale.h>
 #include "lab3.h"
 
-int scalar( const int a[], const int b[], int sz ) {
-    int i = 0; 
+int scalar( const int a[], const int b[], size_t sz ) {
+    size_t i = 0; 
     int ans = 0;
     for ( i = 0; i < sz; i++ ) 
         ans += a[i] * b[i];
